Pulled logic out of MyAccountNavigationController
